# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Media AI Gateway
  version: "2.0.0"
  description: |
    Realtime audio gateway — nhận RTP/WebRTC từ hạ tầng IMS/telecom, xử lý audio,
    gửi sang AI worker (STT/translate) qua gRPC streaming, trả kết quả về backend
    qua HTTP callback hoặc WebRTC DataChannel.

    **DCSF integration (v2):** Gateway tích hợp DCAS — nhận sự kiện cuộc gọi từ DCSF
    tại `/v1/vonras/call-sessions/{callId}/...` thay cho `/v1/sessions` (v1).
servers:
  - url: http://localhost:8080
    description: Control plane (HTTP/2 h2c)
  - url: http://localhost:9090
    description: Metrics server (Prometheus scrape target)
...
paths:
...
  /v1/vonras/call-sessions/{callId}/notify-event:
    post:
      tags: [call-sessions]
      summary: Nhận sự kiện cuộc gọi từ DCSF
      description: |
        DCSF gọi endpoint này để thông báo sự kiện IMS call tới DCAS.

        - **BEGIN** — DCAS ghi nhận cuộc gọi, trả `200 OK` (không tạo session).
        - **ANSWER** — DCAS tạo session audio, cấp RTP port, trả `local_non_dc_media`
          để DCSF dùng làm `remoteNonDcMedia` khi gọi MF MRM `POST /contexts` (3GPP TS29.176).
        - **Service chưa xử lý** (ví dụ `fun_calling`) — DCAS trả `200 OK`, không tạo session.
      operationId: notifyEvent
      parameters:
        - $ref: "#/components/parameters/callId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionEvent"
            examples:
              begin:
                summary: BEGIN event
                value:
                  callIdentifier: "p2uc31@[FC00:0DB8::]"
                  event: "BEGIN"
                  direction: "MT"
                  role: "terminator"
                  bearerCapability: "VIDEO"
                  calling: "86156****5398"
                  called: "86156****5399"
              answer:
                summary: ANSWER event — speech_to_text
                value:
                  callIdentifier: "p2uc31@[FC00:0DB8::]"
                  event: "ANSWER"
                  selectedService: "speech_to_text"
                  direction: "MT"
                  role: "terminator"
                  bearerCapability: "VIDEO"
                  location:
                    areaNumber: "0755"
                    ncgi: "46070094F34050"
                    tac: "000002"
                  calling: "86156****5398"
                  called: "86156****5399"
      responses:
        "200":
          description: |
            BEGIN acknowledged — không tạo session. Cũng trả 200 cho service chưa xử lý.
          content:
            application/json:
              schema:
                type: object
        "201":
          description: ANSWER — session tạo thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
              example:
                session_id: "p2uc31@[FC00:0DB8::]"
                status: "CREATED"
                source_type: "raw_rtp"
                codec: "PCMU"
                sample_rate: 8000
                channels: 1
                task: "speech_to_text"
                created_at: "2026-06-27T08:00:00Z"
                gateway_id: "gw-node-1"
                rtp_ip: "192.168.100.10"
                rtp_port: 40000
                local_non_dc_media:
                  sdpmLine: "audio 40000 RTP/AVP 0"
                  sdpaLines: ["rtpmap:0 PCMU/8000", "ptime:20", "maxptime:20", "recvonly"]
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: callId đã tồn tại (ANSWER đã được xử lý trước đó)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/vonras/call-sessions/{callId}/ctrl-result:
    post:
      tags: [call-sessions]
      summary: Nhận kết quả media negotiation từ DCSF
      description: |
        DCSF forward `ctrl_result` từ IMS-AS sau khi SDP negotiation hoàn tất.
        Gateway cập nhật `remote_addr` (từ `tAccess.endpoint`) cho RTP routing
        và optional `callbackUrl` cho ASR result delivery.

        **DCSF extension:** `callbackUrl` không có trong 3GPP spec — được thêm để
        gateway biết gửi ASR results đến đâu.
      operationId: ctrlResult
      parameters:
        - $ref: "#/components/parameters/callId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CtrlResultRequest"
            example:
              callIdentifier: "p2uc31@[FC00:0DB8::]"
              mediaResources:
                tCore:
                  contextId: "ctx-core-001"
                  terminationId: "term-core-001"
                  endpoint: "10.1.1.1:5060"
                tAccess:
                  contextId: "ctx-access-001"
                  terminationId: "term-access-001"
                  endpoint: "192.168.1.10:5100"
              callbackUrl: "http://mf-host:9090/asr-results"
      responses:
        "200":
          description: Session cập nhật thành công
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /v1/vonras/call-sessions/{callId}:
    get:
      tags: [call-sessions]
      summary: Lấy thông tin session (debug)
      operationId: getCallSession
      parameters:
        - $ref: "#/components/parameters/callId"
      responses:
        "200":
          description: Thông tin session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      tags: [call-sessions]
      summary: Đóng session (RELEASE)
      description: |
        DCSF gọi khi nhận RELEASE event từ IMS-AS.
        Giải phóng pipeline, RTP port, AI stream. Idempotent — 404 nếu không tồn tại.
      operationId: deleteCallSession
      parameters:
        - $ref: "#/components/parameters/callId"
      responses:
        "204":
          description: Session đã được đóng
        "404":
          $ref: "#/components/responses/NotFound"

  # ── WebRTC ────────────────────────────────────────────────────────────────
...
components:
  parameters:
    callId:
      name: callId
      in: path
      required: true
      description: Call session ID — callIdentifier từ DCSF (SIP URI hoặc hash)
      schema:
        type: string
      example: "p2uc31@[FC00:0DB8:0000:0000:0000:0000:0000:0000]"
  responses:
    BadRequest:
      description: Request không hợp lệ
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource không tồn tại
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServiceUnavailable:
      description: Gateway quá tải hoặc không đủ tài nguyên
      headers:
        Retry-After:
          description: Thời gian chờ khuyến nghị (giây)
          schema:
            type: integer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: "service at capacity, try again later"
            retry_after_ms: 5000
  schemas:
...
    MediaResources:
      type: object
      description: Cặp termination H.248 — core-side và access-side.
      properties:
        tCore:
          $ref: "#/components/schemas/MediaResource"
        tAccess:
          $ref: "#/components/schemas/MediaResource"
    EventLocation:
      type: object
      description: Vị trí địa lý của UE trong sự kiện DCSF.
      properties:
        areaNumber:
          type: string
          example: "0755"
        ncgi:
          type: string
          example: "46070094F34050"
        tac:
          type: string
          example: "000002"

    # ── Requests ────────────────────────────────────────────────────────────
    SessionEvent:
      type: object
      required: [callIdentifier, event]
      description: Sự kiện phiên từ DCSF.
      properties:
        callIdentifier:
          type: string
          description: SIP URI định danh cuộc gọi — dùng làm session ID
          example: "p2uc31@[FC00:0DB8:0000:0000:0000:0000:0000:0000]"
        event:
          type: string
          enum: [BEGIN, ANSWER]
          description: Loại sự kiện
          example: "ANSWER"
        selectedService:
          type: string
          description: |
            Dịch vụ đã chọn từ DCSF. Gateway sẽ tạo session cho:
            `speech_to_text`, `realtime_translation` → PCMU/8000.
            Các service khác (ví dụ `fun_calling`) → ACK không tạo session.
          example: "speech_to_text"
        direction:
          type: string
          enum: [MO, MT]
          example: "MT"
        role:
          type: string
          enum: [originator, terminator]
          example: "terminator"
        bearerCapability:
          type: string
          enum: [AUDIO, VIDEO]
          description: Loại bearer. Cả AUDIO và VIDEO đều dùng PCMU/8000 cho STT.
          example: "VIDEO"
        location:
          $ref: "#/components/schemas/EventLocation"
        calling:
          type: string
          description: MSISDN của bên gọi
          example: "86156****5398"
        called:
          type: string
          description: MSISDN của bên nhận
          example: "86156****5399"
    CtrlResultRequest:
      type: object
      required: [callIdentifier]
      description: |
        Ctrl result từ IMS-AS forwarded bởi DCSF sau SDP negotiation.
        `callbackUrl` là DCSF extension — không có trong 3GPP spec.
      properties:
        callIdentifier:
          type: string
          description: SIP URI định danh cuộc gọi — phải khớp với session ID đã tạo
          example: "p2uc31@[FC00:0DB8::]"
        actionResults:
          type: array
          description: Kết quả xử lý từ IMS-AS. Gateway không xử lý, chỉ forward-compatible.
          items:
            type: object
        mediaResources:
          $ref: "#/components/schemas/MediaResources"
        callbackUrl:
          type: string
          format: uri
          description: |
            HTTP/2 endpoint nhận ASR results (DCSF extension).
            `tAccess.endpoint` tự động được set làm `remote_addr` cho RTP routing.
          example: "http://mf-host:9090/asr-results"
...
    SessionResponse:
      type: object
      properties:
        session_id:
          type: string
          description: Session ID (= callId = callIdentifier từ DCSF)
          example: "p2uc31@[FC00:0DB8::]"
        status:
          type: string
          enum: [CREATED, ACTIVE, CLOSED, ERROR]
          example: "CREATED"
        source_type:
          type: string
          enum: [raw_rtp, webrtc]
          example: "raw_rtp"
        codec:
          type: string
          example: "PCMU"
        sample_rate:
          type: integer
          example: 8000
        channels:
          type: integer
          example: 1
        language:
          type: string
          example: "vi"
        task:
          type: string
          description: selectedService từ DCSF (ví dụ speech_to_text, realtime_translation)
          example: "speech_to_text"
        created_at:
          type: string
          format: date-time
          example: "2026-06-27T08:00:00Z"
        gateway_id:
          type: string
          description: Node identity (chỉ có trong ANSWER CREATE response)
          example: "gw-node-1"
        rtp_ip:
          type: string
          description: Public IP để gửi RTP vào (chỉ có trong ANSWER CREATE response)
          example: "192.168.100.10"
        rtp_port:
          type: integer
          description: UDP port để gửi RTP vào (chỉ có trong ANSWER CREATE response)
          example: 40000
        local_non_dc_media:
          $ref: "#/components/schemas/NonDcMedia"
          description: |
            SDP mô tả gateway RTP endpoint (chỉ có trong ANSWER CREATE response, raw_rtp với per-session port pool).
            DCSF dùng làm remoteNonDcMedia khi gọi MF MRM POST /contexts (3GPP TS29.176).
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "session not found"
        retry_after_ms:
          type: integer
          format: int64
          description: Thời gian chờ khuyến nghị trước khi retry (ms). 0 = không retry.
          example: 5000